stmmac: document snps,eee-force-disable DT parameter
This patch is to document the eee-force-disable device tree parameter
that can be passed to the stmmac for disabling the EEE at phy level.

This option should be considered a temporary solution because in the
new kernels it should be moved on PHY and managed by using
of_set_phy_supported (as max-speed).

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Alexandre TORGUE <alexandre.torgue@st.com>
Signed-off-by: Tomer Maimon <tmaimon77@gmail.com>